Move digital-funding to digital-fund
