+++ controllers/digital-fund/views/full-eligibility.njk
@@ -61,7 +61,7 @@
                     <li>Products or services that duplicate existing services with no substantial innovation</li>
                 </ul>
 
-                <p><a href="/funding/programmes/digital-funding/strand-1/eligibility"
+                <p><a href="/funding/programmes/digital-fund/strand-1/eligibility"
                     class="btn btn--medium accent--{{ pageAccent }} a--btn u-margin-bottom">Request a call about strand one</a></p>
 
 
@@ -87,7 +87,7 @@
                     <li>Proposals that are based around core budget shortfalls.</li>
                 </ul>
 
-                <p><a href="/funding/programmes/digital-funding/strand-2/eligibility"
+                <p><a href="/funding/programmes/digital-fund/strand-2/eligibility"
                     class="btn btn--medium accent--{{ pageAccent }} a--btn">Request a call about strand two</a></p>
             {% endcall %}
 
@@ -167,7 +167,7 @@
                             <div class="card__body">
                                 <p>Grants of up to £500,000, plus a package of tailored support, to help established established charities that want to use digital to help take a major leap forward</p>
 
-                                <p><a href="/funding/programmes/digital-funding/strand-1"
+                                <p><a href="/funding/programmes/digital-fund/strand-1"
                                     class="btn btn--medium accent--blue a--btn">
                                     Find out more about this strand
                                 </a></p>
@@ -183,7 +183,7 @@
                                 <p>Grants of up to £500,000 for tech for good organisations that have already launched promising services</p>
 
 
-                                <p><a href="/funding/programmes/digital-funding/strand-2"
+                                <p><a href="/funding/programmes/digital-fund/strand-2"
                                     class="btn btn--medium accent--cyan a--btn">
                                     Find out more about this strand
                                 </a></p>
