+++ controllers/digital-fund/views/landing.njk
@@ -38,7 +38,7 @@
                             title: copy.strand1.title,
                             description: copy.strand1.teaserDescription,
                             link: {
-                                url: localify('/funding/programmes/digital-funding/strand-1'),
+                                url: localify('/funding/programmes/digital-fund/strand-1'),
                                 label: copy.aboutStrandLabel
                             }
                         }, accent = 'blue') }}
@@ -48,7 +48,7 @@
                             title: copy.strand2.title,
                             description: copy.strand2.teaserDescription,
                             link: {
-                                url: localify('/funding/programmes/digital-funding/strand-2'),
+                                url: localify('/funding/programmes/digital-fund/strand-2'),
                                 label: copy.aboutStrandLabel
                             }
                         }, accent = 'cyan') }}
@@ -57,7 +57,7 @@
 
                 <p>
                     {{ copy.getStarted.fullEligiblity.prefix }}
-                    <a href="{{ localify('/funding/programmes/digital-funding/eligibility') }}">
+                    <a href="{{ localify('/funding/programmes/digital-fund/eligibility') }}">
                         {{ copy.getStarted.fullEligiblity.linkText }}
                     </a>
                 </p>
