+++ controllers/digital-fund/views/strand-2-eligibility.njk
@@ -23,11 +23,11 @@
 
                     <div class="o-button-group u-margin-bottom">
                         <a class="btn btn--medium accent--{{ pageAccent }} a--btn"
-                           href="{{ localify('/apply/digital-funding-2/1') }}">
+                           href="{{ localify('/apply/digital-fund-strand-2/1') }}">
                             {{ __('global.misc.yes') }}
                         </a>
                         <a class="btn btn--medium accent--{{ pageAccent }} a--btn"
-                           href="{{ localify('/funding/programmes/digital-funding/strand-2/eligibility/ineligible') }}">
+                           href="{{ localify('/funding/programmes/digital-fund/strand-2/eligibility/ineligible') }}">
                             {{ __('global.misc.no') }}
                         </a>
                     </div>
@@ -42,7 +42,7 @@
                                         <li>{{ criterion }}</li>
                                     {% endfor %}
                                 </ul>
-                                {% set readMoreLink = localify('/funding/programmes/digital-funding/strand-2#eligibility') %}
+                                {% set readMoreLink = localify('/funding/programmes/digital-fund/strand-2#eligibility') %}
                                 <p>{{ __('funding.digitalFunding.strand2.eligibilityDetail.unsure.readMore', readMoreLink) | safe }}</p>
                             </div>
                         </div>
