+++ controllers/digital-fund/views/view-helpers.njk
@@ -18,19 +18,19 @@
 {% macro wayfinding(currentStrand) %}
     <h2>{{ copy.wayfinding.title }}</h2>
     {% if currentStrand !== 'assistance' %}
-        <p><a href="{{ localify('/funding/programmes/digital-funding/assistance') }}"
+        <p><a href="{{ localify('/funding/programmes/digital-fund/assistance') }}"
             class="btn btn--medium accent--pink a--btn">
             {{ copy.wayfinding.assistance }}
         </a></p>
     {% endif %}
     {% if currentStrand !== 'strand1' %}
-        <p><a href="{{ localify('/funding/programmes/digital-funding/strand-1') }}"
+        <p><a href="{{ localify('/funding/programmes/digital-fund/strand-1') }}"
             class="btn btn--medium accent--blue a--btn">
             {{ copy.wayfinding.strand1 }}
         </a></p>
     {% endif %}
     {% if currentStrand !== 'strand2' %}
-        <p><a href="{{ localify('/funding/programmes/digital-funding/strand-2') }}"
+        <p><a href="{{ localify('/funding/programmes/digital-fund/strand-2') }}"
             class="btn btn--medium accent--cyan a--btn">
             {{ copy.wayfinding.strand2 }}
         </a></p>
